backend/requirements.txt
Django==4.2.11
psycopg2-binary>=2.9.9
gunicorn==20.1.0
python-dotenv==0.21.0
requests==2.31.0
cryptography==42.0.7
certifi==2024.2.2
PyYAML>=6.0,<7.0
idna==3.7
django-browser-reload==1.9.0
django-hcaptcha
whitenoise[brotli]
# Пакеты для блога
django-parler>=2.3          # Многоязычность
django-ckeditor>=6.7.0      # Богатый текстовый редактор  
django-taggit>=4.0.0        # Система тегов
django-meta>=2.2.0          # SEO мета-теги
sorl-thumbnail>=12.10.0     # Обработка изображений
Pillow>=10.0.0              # Работа с изображениями
redis>=4.5.0                # Кэширование
django-redis>=5.3.0         # Redis для Django